Aprende c flex

+++ scripts/clear_temp.sh
@@ -1,4 +1,20 @@
 #!/bin/bash
 cd /home/rodrigo/Documents/Compiladores/tmp
-echo Limpando Temp
-rm -f *
+while [ true ]
+do
+  read -p "Queres realmente limpar /tmp [y/n]? " RESPOSTA
+  case $RESPOSTA in
+    y)
+      echo Limpando Temp
+      rm -f *
+      break
+      ;;
+    n)
+      echo Cancelado
+      break
+      ;;
+    *)
+      echo Por favor digite \'y\' ou \'n\'
+      ;;
+  esac
+done
